moved math implementation of cross entropy loss from ml to math
